backend/requirements.txt
acme==2.6.0
annotated-types==0.7.0
anyio==4.9.0
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.25.0
awscrt==0.23.8
Babel==2.9.1
certbot==2.6.0
certbot-nginx==2.6.0
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
click==8.1.8
cloud-init==22.2.2
colorama==0.4.4
ConfigArgParse==1.7
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.8
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
gpg==1.23.2
h11==0.16.0
idna==2.10
Jinja2==2.11.3
jmespath==0.10.0
josepy==1.13.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==1.1.1
netifaces==0.10.6
oauthlib==3.0.2
parsedatetime==2.6
passlib==1.7.4
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.20
pydantic==2.11.7
pydantic_core==2.33.2
pyOpenSSL==21.0.0
pyparsing>=3.0.4
pyRFC3339==1.1
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.32.3
rpm==4.16.1.3
rsa==4.9.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
sniffio==1.3.1
starlette==0.47.2
support-info==1.0
systemd-python==235
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.25.10
wcwidth==0.2.5
boto3>=1.35.23
pydantic-settings>=2.5.2
orjson>=3.10   # faster JSON responses (FastAPI will auto-use it if installed)
uvicorn[standard]==0.30.6

google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-oauthlib>=1.0.0
anthropic>=0.40.0
Pillow>=11.0.0
PyMuPDF>=1.24.0
ffmpeg-python>=0.2.0
librosa>=0.10.0
matplotlib>=3.7.0
numpy>=1.24.0
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0  # Celery monitoring
